调整条码解析接口、库存任务服务

+++ src/main/java/org/edi/stocktask/bo/stocktask/IStockTaskItem.java
@@ -1,5 +1,7 @@
 package org.edi.stocktask.bo.stocktask;
 
+import java.math.BigDecimal;
+
 /**
  * @author Fancy
  * @date 2018/5/25
@@ -58,9 +60,9 @@ public interface IStockTaskItem {
 
     void setQuantity(Double value);
 
-    BigDecimal getOpenQuantity();
+    Double getOpenQuantity();
 
-    void setOpenQuantity(BigDecimal value);
+    void setOpenQuantity(Double value);
 
     String getInventoryUoM();
